+++ blender-qmm/mats/extras/Cinnabar.py
@@ -7,28 +7,28 @@ def ShowMessageBox(message = "", title = "", icon = 'INFO'):
         self.layout.label(text=message)
     bpy.context.window_manager.popup_menu(draw, title = title, icon = icon)
 
-#Red MetalShaderOperator
-class QMMRedMetal(bpy.types.Operator):
-    """Add/Apply Red Metal Material to Selected Object (or Scene)"""
-    bl_label = "QMM Red Metal Shader"
-    bl_idname = 'shader.qmm_red_metal_operator'
+#CinnabarShaderOperator
+class QMMCinnabar(bpy.types.Operator):
+    """Add/Apply Cinnabar Material to Selected Object (or Scene)"""
+    bl_label = "QMM Cinnabar Shader"
+    bl_idname = 'shader.qmm_cinnabar_operator'
     def execute(self, context):
         # DOES THE MATERIAL ALREADY EXIST?
-        m_red_metal = bpy.data.materials.get("QMM Red Metal")
-        if m_red_metal:
-            ShowMessageBox(message_text, "QMM Red Metal")
-            # print(f"QMM Red Metal already exists")
-            bpy.context.object.active_material = m_red_metal
+        m_cinnabar = bpy.data.materials.get("QMM Cinnabar")
+        if m_cinnabar:
+            ShowMessageBox(message_text, "QMM Cinnabar")
+            # print(f"QMM Cinnabar already exists")
+            bpy.context.object.active_material = m_cinnabar
             return {'FINISHED'}
         else:
             #CreateShader
-            m_red_metal = bpy.data.materials.new(name = "QMM Red Metal")
-            m_red_metal.use_nodes = True
-            m_red_metal.diffuse_color = (0.768151, 0.054480, 0.034340, 1)
-            m_red_metal.metallic = 1
-            m_red_metal.roughness = 0.6
+            m_cinnabar = bpy.data.materials.new(name = "QMM Cinnabar")
+            m_cinnabar.use_nodes = True
+            m_cinnabar.diffuse_color = (0.768151, 0.054480, 0.034340, 1)
+            m_cinnabar.metallic = 1
+            m_cinnabar.roughness = 0.6
 
-            nodes = m_red_metal.node_tree.nodes
+            nodes = m_cinnabar.node_tree.nodes
 
             #materialoutput
             material_output = nodes.get('Material Output')
@@ -40,14 +40,14 @@ class QMMRedMetal(bpy.types.Operator):
             BSDF.inputs[0].default_value = (0.768151, 0.054480, 0.034340, 1)
             BSDF.inputs[6].default_value = 1
             BSDF.inputs[9].default_value = 0.4
-            BSDF.inputs[16].default_value = 3.256
+            BSDF.inputs[16].default_value = 3.02
 
             #colorramp
             m_colorramp = nodes.new('ShaderNodeValToRGB')
-            m_colorramp.label = "Red Metal"
+            m_colorramp.label = "Cinnabar"
             m_colorramp.location = (-800, -200)
             m_colorramp.color_ramp.elements[0].color = (0.768151, 0.054480, 0.034340, 1)
-            m_colorramp.color_ramp.elements[1].color = (0.964687, 0.066626, 0.002428, 1)
+            m_colorramp.color_ramp.elements[1].color = (0.283149, 0.039546, 0.031896, 1)
             m_colorramp.color_ramp.elements[0].position = 0.2
             m_colorramp.color_ramp.elements[1].position = 0.6
 
@@ -60,7 +60,7 @@ class QMMRedMetal(bpy.types.Operator):
             m_colorramp2.color_ramp.elements[0].position = 0.2
             m_colorramp2.color_ramp.elements[1].position = 0.6
 
-            nodes = m_red_metal.node_tree.nodes
+            nodes = m_cinnabar.node_tree.nodes
 
             #EnergyConservationGroup
             bpy.ops.node.ec_group_operator()
@@ -68,7 +68,7 @@ class QMMRedMetal(bpy.types.Operator):
             ec_group.name = "Energy Conservation"
             ec_group.node_tree = bpy.data.node_groups['Energy Conservation']
             ec_group.location = (-500, -200)
-            ec_group.inputs[0].default_value = 3.256
+            ec_group.inputs[0].default_value = 3.02
             ec_group.inputs[1].default_value = (0.768151, 0.054480, 0.034340, 1)
             ec_group.inputs[2].default_value = (0.010000, 0.000687, 0.000429, 1)
 
@@ -83,7 +83,7 @@ class QMMRedMetal(bpy.types.Operator):
             texturizer_group.inputs[4].default_value = 0.0125
 
             #Links
-            links = m_red_metal.node_tree.links.new
+            links = m_cinnabar.node_tree.links.new
 
             links(m_colorramp.outputs[0], ec_group.inputs[1])
             links(texturizer_group.outputs[1], m_colorramp.inputs[0])
@@ -94,6 +94,6 @@ class QMMRedMetal(bpy.types.Operator):
             links(ec_group.outputs[3], BSDF.inputs[16])
 
             #LOAD THE MATERIAL
-            bpy.context.object.active_material = m_red_metal
+            bpy.context.object.active_material = m_cinnabar
 
         return {'FINISHED'}