+++ blender-qmm/mats/noble/FresnelSilver.py
@@ -72,6 +72,7 @@ class QMMSilverFresnel(bpy.types.Operator):
             bpy.ops.node.silver_colors_group_operator()
             nodes = m_silver.node_tree.nodes
             silver_colors_group = nodes.new("ShaderNodeGroup")
+            silver_colors_group.name = "Silver Colors"
             silver_colors_group.node_tree = bpy.data.node_groups['Silver Colors']
             silver_colors_group.location = (-600, 0)
             links(silver_colors_group.outputs[0], m_glossy.inputs[0])